+++ R/berkeley_exdata.R
@@ -3,13 +3,12 @@
 #' @title Berkeley Child Guidance Study Data for males
 #'
 #' @description A subset of the [bsitar::berkeley] data that contains
-#'   longitudinal growth data for 20 randomly selected females (8 to 18
-#'   years of age).
+#'   longitudinal growth data for 70 females (8 to 18 years of age).
 #'
 #' @details A detailed description of the full data is provided in the
 #'   [bsitar::berkeley] data.
 #'
-#' @name berkeley_fdata
+#' @name berkeley_exdata
 #' @docType data
 #' @format A data frame with following 3 variables:
 #' \describe{
@@ -22,4 +21,4 @@
 #'
 #' @return A data frame with 3 columns.
 #' @inherit berkeley author
-"berkeley_fdata"
+"berkeley_exdata"