+++ R/utils-helper-1.R
@@ -2446,3 +2446,247 @@ add_parms_to_curve_data <- function(data,
 
 
 
+#' An internal function to bind rows of unequal lengths (adapted from qpcR:::cbind.na)
+#'
+#' @param deparse.level An integer to set deparse level. 
+#' @param ... A list or name of column vectors.
+#' @keywords internal
+#' @return A data frame.
+#' @keywords internal
+#' @noRd
+#'
+rbind_fill_na1 <- function (..., deparse.level = 1) {
+    na <- nargs() - (!missing(deparse.level))
+    deparse.level <- as.integer(deparse.level)
+    stopifnot(0 <= deparse.level, deparse.level <= 2)
+    argl <- list(...)
+    while (na > 0 && is.null(argl[[na]])) {
+      argl <- argl[-na]
+      na <- na - 1
+    }
+    if (na == 0) 
+      return(NULL)
+    if (na == 1) {
+      if (isS4(..1)) 
+        return(methods::rbind2(..1))
+      else return(matrix(..., nrow = 1))
+    }
+    if (deparse.level) {
+      symarg <- as.list(sys.call()[-1L])[1L:na]
+      Nms <- function(i) {
+        if (is.null(r <- names(symarg[i])) || r == "") {
+          if (is.symbol(r <- symarg[[i]]) || deparse.level == 
+              2) 
+            deparse(r)
+        }
+        else r
+      }
+    }
+    if (na == 0) {
+      r <- argl[[2]]
+      fix.na <- FALSE
+    }
+    else {
+      nrs <- unname(lapply(argl, ncol))
+      iV <- sapply(nrs, is.null)
+      fix.na <- identical(nrs[(na - 1):na], list(NULL, NULL))
+      if (deparse.level) {
+        if (fix.na) 
+          fix.na <- !is.null(Nna <- Nms(na))
+        if (!is.null(nmi <- names(argl))) 
+          iV <- iV & (nmi == "")
+        ii <- if (fix.na) 
+          2:(na - 1)
+        else 2:na
+        if (any(iV[ii])) {
+          for (i in ii[iV[ii]]) if (!is.null(nmi <- Nms(i))) 
+            names(argl)[i] <- nmi
+        }
+      }
+      nCol <- as.numeric(sapply(argl, function(x) if (is.null(ncol(x))) length(x) else ncol(x)))
+      maxCol <- max(nCol, na.rm = TRUE)
+      argl <- lapply(argl, function(x) if (is.null(ncol(x))) 
+        c(x, rep(NA, maxCol - length(x)))
+        else cbind(x, matrix(, nrow(x), maxCol - ncol(x))))
+      namesVEC <- rep(NA, maxCol)
+      for (i in 1:length(argl)) {
+        CN <- colnames(argl[[i]])
+        m <- !(CN %in% namesVEC)
+        namesVEC[m] <- CN[m]
+      }
+      for (j in 1:length(argl)) {
+        if (!is.null(ncol(argl[[j]]))) 
+          colnames(argl[[j]]) <- namesVEC
+      }
+      r <- do.call(rbind, c(argl[-1L], list(deparse.level = deparse.level)))
+    }
+    d2 <- dim(r)
+    colnames(r) <- colnames(argl[[1]])
+    r <- methods::rbind2(argl[[1]], r)
+    if (deparse.level == 0) 
+      return(r)
+    ism1 <- !is.null(d1 <- dim(..1)) && length(d1) == 2L
+    ism2 <- !is.null(d2) && length(d2) == 2L && !fix.na
+    if (ism1 && ism2) 
+      return(r)
+    Nrow <- function(x) {
+      d <- dim(x)
+      if (length(d) == 2L) 
+        d[1L]
+      else as.integer(length(x) > 0L)
+    }
+    nn1 <- !is.null(N1 <- if ((l1 <- Nrow(..1)) && !ism1) Nms(1))
+    nn2 <- !is.null(N2 <- if (na == 2 && Nrow(..2) && !ism2) Nms(2))
+    if (nn1 || nn2 || fix.na) {
+      if (is.null(rownames(r))) 
+        rownames(r) <- rep.int("", nrow(r))
+      setN <- function(i, nams) rownames(r)[i] <<- if (is.null(nams)) 
+        ""
+      else nams
+      if (nn1) 
+        setN(1, N1)
+      if (nn2) 
+        setN(1 + l1, N2)
+      if (fix.na) 
+        setN(nrow(r), Nna)
+    }
+    r
+  }
+
+
+#' An internal function to bind columns of unequal lengths (adapted from qpcR:::cbind.na)
+#'
+#' @param deparse.level An integer to set deparse level. 
+#' @param ... A list or name of column vectors.
+#' @keywords internal
+#' @return A data frame.
+#' @keywords internal
+#' @noRd
+#'
+cbind_fill_na1 <- function (..., deparse.level = 1) {
+  na <- nargs() - (!missing(deparse.level))
+  deparse.level <- as.integer(deparse.level)
+  stopifnot(0 <= deparse.level, deparse.level <= 2)
+  argl <- list(...)
+  while (na > 0 && is.null(argl[[na]])) {
+    argl <- argl[-na]
+    na <- na - 1
+  }
+  if (na == 0) 
+    return(NULL)
+  if (na == 1) {
+    if (isS4(..1)) 
+      return(methods::cbind2(..1))
+    else return(matrix(...))
+  }
+  if (deparse.level) {
+    symarg <- as.list(sys.call()[-1L])[1L:na]
+    Nms <- function(i) {
+      if (is.null(r <- names(symarg[i])) || r == "") {
+        if (is.symbol(r <- symarg[[i]]) || deparse.level == 
+            2) 
+          deparse(r)
+      }
+      else r
+    }
+  }
+  if (na == 0) {
+    r <- argl[[2]]
+    fix.na <- FALSE
+  }
+  else {
+    nrs <- unname(lapply(argl, nrow))
+    iV <- sapply(nrs, is.null)
+    fix.na <- identical(nrs[(na - 1):na], list(NULL, NULL))
+    if (deparse.level) {
+      if (fix.na) 
+        fix.na <- !is.null(Nna <- Nms(na))
+      if (!is.null(nmi <- names(argl))) 
+        iV <- iV & (nmi == "")
+      ii <- if (fix.na) 
+        2:(na - 1)
+      else 2:na
+      if (any(iV[ii])) {
+        for (i in ii[iV[ii]]) if (!is.null(nmi <- Nms(i))) 
+          names(argl)[i] <- nmi
+      }
+    }
+    nRow <- as.numeric(sapply(argl, function(x) NROW(x)))
+    maxRow <- max(nRow, na.rm = TRUE)
+    argl <- lapply(argl, function(x) if (is.null(nrow(x))) 
+      c(x, rep(NA, maxRow - length(x)))
+      else rbind_fill_na1(x, matrix(, maxRow - nrow(x), ncol(x))))
+    r <- do.call(cbind, c(argl[-1L], list(deparse.level = deparse.level)))
+  }
+  d2 <- dim(r)
+  r <- methods::cbind2(argl[[1]], r)
+  if (deparse.level == 0) 
+    return(r)
+  ism1 <- !is.null(d1 <- dim(..1)) && length(d1) == 2L
+  ism2 <- !is.null(d2) && length(d2) == 2L && !fix.na
+  if (ism1 && ism2) 
+    return(r)
+  Ncol <- function(x) {
+    d <- dim(x)
+    if (length(d) == 2L) 
+      d[2L]
+    else as.integer(length(x) > 0L)
+  }
+  nn1 <- !is.null(N1 <- if ((l1 <- Ncol(..1)) && !ism1) Nms(1))
+  nn2 <- !is.null(N2 <- if (na == 2 && Ncol(..2) && !ism2) Nms(2))
+  if (nn1 || nn2 || fix.na) {
+    if (is.null(colnames(r))) 
+      colnames(r) <- rep.int("", ncol(r))
+    setN <- function(i, nams) colnames(r)[i] <<- if (is.null(nams)) 
+      ""
+    else nams
+    if (nn1) 
+      setN(1, N1)
+    if (nn2) 
+      setN(1 + l1, N2)
+    if (fix.na) 
+      setN(ncol(r), Nna)
+  }
+  r
+}
+
+
+
+
+
+#' An internal function to bind columns of unequal lengths
+#'
+#' @param names A vector of character string to name columns. 
+#' @param ... A list or names of column vectors.
+#' @keywords internal
+#' @return A data frame.
+#' @keywords internal
+#' @noRd
+#'
+cbind_fill_na2 <- function(..., names = NA) {
+  xlist = list(...)
+  cbindfill.id <- NULL;
+  suppressWarnings({
+    y= Reduce(
+      function(a,b) {
+        if(is.vector(a)) na = length(a)
+        if(is.factor(a)) na = levels(a)
+        if(is.factor(a)) a = data.frame(a) %>% droplevels()
+        if(is.data.frame(a)|is.matrix(a)) na = nrow(a)
+        if(is.vector(b)) nb = length(b)
+        if(is.factor(b)) nb = levels(b)
+        if(is.factor(b)) b = data.frame(b) %>% droplevels()
+        if(is.data.frame(b)|is.matrix(b)) nb = nrow(b)
+        subset(
+          merge(
+            cbind(cbindfill.id = 1:na, a),
+            cbind(cbindfill.id = 1:nb, b),
+            all = TRUE, by = "cbindfill.id"
+          ),
+          select = -cbindfill.id
+        )}
+      ,xlist)
+    if(is.na(names)) colnames(y) <- names(xlist) else colnames(y) <- names
+  })
+  return(y)
+}